+++ config/seata-config-core/src/main/java/io/seata/config/ConfigurationKeys.java
@@ -41,4 +41,8 @@ public interface ConfigurationKeys {
      * The constant FILE_ROOT_TYPE.
      */
     String FILE_ROOT_TYPE = "type";
+    /**
+     * The constant DATA_TYPE.
+     */
+    String DATA_TYPE = "dataType";
 }
