+++ seata-spring-autoconfigure/seata-spring-autoconfigure-core/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -97,18 +97,7 @@
       ]
     },
     {
-      "name": "seata.config.nacos.data-type",
-      "providers": [
-        {
-          "name": "handle-as",
-          "parameters": {
-            "target": "io.seata.config.processor.ConfigDataType"
-          }
-        }
-      ]
-    },
-    {
-      "name": "seata.config.zk.data-type",
+      "name": "seata.config.data-type",
       "providers": [
         {
           "name": "handle-as",
